pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uberscrape"
version = "0.1.0"
description = "AI-powered web scraping that returns structured data, not HTML walls"
readme = "README.md"
authors = [{name = "Doug Simpson", email = "dougiefreshcodes@gmail.com"}]
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "httpx>=0.28.0",
    "playwright>=1.58.0",
    "beautifulsoup4>=4.14.0",
    "html2text>=2025.4.0",
    "anthropic>=0.84.0",
    "click>=8.1.0",
    "rich>=14.3.0",
    "pydantic>=2.12.0",
    "json-repair>=0.44.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
uberscrape = "uberscrape.cli:cli"

[project.urls]
Homepage = "https://github.com/dougsimpsoncodes/uberscrape"
Repository = "https://github.com/dougsimpsoncodes/uberscrape"
Issues = "https://github.com/dougsimpsoncodes/uberscrape/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["uberscrape*"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"
